# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  id: number;
  label: string;
  loginid: string;
  /** real or demo */
  accountType: string;
  currency: string;
  balance: number;
  email?: string;
  country?: string;
  isActive: boolean;
  connectedAt: string;
}

export interface ConnectAccountBody {
  /** @minLength 1 */
  label: string;
  /**
   * Deriv API token (read/trade scopes)
   * @minLength 8
   */
  apiToken: string;
}

export interface UpdateAccountBody {
  label?: string;
  isActive?: boolean;
}

export type BotStatus = (typeof BotStatus)[keyof typeof BotStatus];

export const BotStatus = {
  idle: "idle",
  running: "running",
  paused: "paused",
} as const;

export interface Bot {
  id: number;
  name: string;
  description?: string;
  /** e.g. Martingale, D'Alembert, Custom */
  strategy: string;
  /** e.g. Volatility 75, Boom 1000 */
  market: string;
  tags: string[];
  favorite: boolean;
  sizeBytes: number;
  status: BotStatus;
  lastRunAt?: string | null;
  createdAt: string;
}

export type BotDetail = Bot & {
  /** Full Deriv DBot XML content */
  xmlContent: string;
};

export interface ImportBotBody {
  /** @minLength 1 */
  name: string;
  description?: string;
  strategy?: string;
  market?: string;
  tags?: string[];
  /** @minLength 1 */
  xmlContent: string;
}

export type UpdateBotBodyStatus =
  (typeof UpdateBotBodyStatus)[keyof typeof UpdateBotBodyStatus];

export const UpdateBotBodyStatus = {
  idle: "idle",
  running: "running",
  paused: "paused",
} as const;

export interface UpdateBotBody {
  name?: string;
  description?: string;
  strategy?: string;
  market?: string;
  tags?: string[];
  status?: UpdateBotBodyStatus;
}

export interface Tool {
  id: string;
  name: string;
  description: string;
  category: string;
  icon: string;
}

export interface MartingaleBody {
  /** @minimum 0.01 */
  baseStake: number;
  /** @minimum 1 */
  multiplier: number;
  /**
   * @minimum 1
   * @maximum 25
   */
  steps: number;
  /** e.g. 95 for 95% payout */
  payoutPercent: number;
}

export interface MartingaleStep {
  step: number;
  stake: number;
  cumulativeLoss: number;
  payout: number;
  netIfWin: number;
}

export interface MartingaleResult {
  steps: MartingaleStep[];
  totalRisk: number;
}

export interface RiskBody {
  /** @minimum 0 */
  accountBalance: number;
  /**
   * @minimum 0.1
   * @maximum 100
   */
  riskPercent: number;
  /** @minimum 1 */
  stopLossPips: number;
  /** @minimum 0.01 */
  pipValue: number;
}

export interface RiskResult {
  riskAmount: number;
  positionSize: number;
}

export interface CompoundBody {
  /** @minimum 0 */
  startingBalance: number;
  dailyReturnPercent: number;
  /**
   * @minimum 1
   * @maximum 365
   */
  days: number;
}

export interface CompoundPoint {
  day: number;
  balance: number;
}

export interface CompoundResult {
  points: CompoundPoint[];
  finalBalance: number;
  totalProfit: number;
}

export type TradeResult = (typeof TradeResult)[keyof typeof TradeResult];

export const TradeResult = {
  win: "win",
  loss: "loss",
  breakeven: "breakeven",
} as const;

export interface Trade {
  id: number;
  symbol: string;
  /** e.g. CALL, PUT, RISE, FALL, DIGITS */
  contractType: string;
  stake: number;
  payout: number;
  profit: number;
  result: TradeResult;
  botId?: number | null;
  notes?: string;
  tradedAt: string;
}

export type CreateTradeBodyResult =
  (typeof CreateTradeBodyResult)[keyof typeof CreateTradeBodyResult];

export const CreateTradeBodyResult = {
  win: "win",
  loss: "loss",
  breakeven: "breakeven",
} as const;

export interface CreateTradeBody {
  /** @minLength 1 */
  symbol: string;
  /** @minLength 1 */
  contractType: string;
  stake: number;
  payout: number;
  profit: number;
  result: CreateTradeBodyResult;
  botId?: number | null;
  notes?: string;
  tradedAt?: string;
}

export interface DashboardSummary {
  connectedAccounts: number;
  totalBots: number;
  runningBots: number;
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  totalProfit: number;
  bestSymbol?: string | null;
  activeAccount?: Account | null;
}

export interface EquityPoint {
  date: string;
  equity: number;
}

export interface SymbolBreakdownItem {
  symbol: string;
  trades: number;
  profit: number;
  winRate: number;
}

export type ListBotsParams = {
  /**
   * Search by name or tag
   */
  q?: string;
};

export type ListTradesParams = {
  limit?: number;
  result?: ListTradesResult;
};

export type ListTradesResult =
  (typeof ListTradesResult)[keyof typeof ListTradesResult];

export const ListTradesResult = {
  win: "win",
  loss: "loss",
  breakeven: "breakeven",
} as const;