# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeHub - Deriv Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List connected Deriv accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: connectAccount
      tags: [accounts]
      summary: Connect a Deriv account using an API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectAccountBody"
      responses:
        "201":
          description: Account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    delete:
      operationId: disconnectAccount
      tags: [accounts]
      summary: Disconnect (remove) an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account (set as active, rename label)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        loginid:
          type: string
        accountType:
          type: string
          description: real or demo
        currency:
          type: string
        balance:
          type: number
        email:
          type: string
        country:
          type: string
        isActive:
          type: boolean
        connectedAt:
          type: string
          format: date-time
      required: [id, label, loginid, accountType, currency, balance, isActive, connectedAt]
    ConnectAccountBody:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        apiToken:
          type: string
          minLength: 8
          description: Deriv API token (read/trade scopes)
      required: [label, apiToken]
    UpdateAccountBody:
      type: object
      properties:
        label:
          type: string
        isActive:
          type: boolean
...